# docs/api-specs/pipeline.openapi.json
{
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    }
  },
  "info": {
    "description": "File upload, column detection, data cleaning, and analysis pipeline",
    "title": "Kaori Data Pipeline",
    "version": "1.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/pipelines": {
      "get": {
        "description": "List pipeline runs for the calling tenant, newest first.",
        "operationId": "list_pipelines_pipelines_get",
        "parameters": [
          {
            "description": "Opaque cursor from prior meta.cursor",
            "in": "query",
            "name": "cursor",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Opaque cursor from prior meta.cursor",
              "title": "Cursor"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 50,
              "maximum": 500,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "description": "Comma-separated status filter",
            "in": "query",
            "name": "status",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Comma-separated status filter",
              "title": "Status"
            }
          },
          {
            "description": "ISO-8601 lower bound on created_at",
            "in": "query",
            "name": "from",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "date-time",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "ISO-8601 lower bound on created_at",
              "title": "From"
            }
          },
          {
            "description": "ISO-8601 upper bound on created_at",
            "in": "query",
            "name": "to",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "date-time",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "ISO-8601 upper bound on created_at",
              "title": "To"
            }
          },
          {
            "in": "header",
            "name": "X-Enterprise-ID",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "X-Enterprise-Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Pipelines",
        "tags": [
          "Pipelines"
        ]
      }
    },
    "/pipelines/{run_id}/events": {
      "get": {
        "description": "Server-Sent Events for a single pipeline run.\n\nEmits one event per ``pipeline_runs.status`` transition for the run.\nHeartbeats every 15 s as comment lines so reverse proxies (nginx,\nAWS ALB) don't close the idle TCP connection. Browsers reconnect\nautomatically with ``Last-Event-ID`` — when present, we replay the\ncurrent row state immediately so the client sees the latest known\nstatus without having to wait for the next transition.",
        "operationId": "status_stream_pipelines__run_id__events_get",
        "parameters": [
          {
            "description": "Pipeline run UUID",
            "in": "path",
            "name": "run_id",
            "required": true,
            "schema": {
              "description": "Pipeline run UUID",
              "format": "uuid",
              "title": "Run Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "X-Enterprise-ID",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "X-Enterprise-Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "Last-Event-ID",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Last-Event-Id"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Status Stream",
        "tags": [
          "Pipelines"
        ]
      }
    },
...
  }
}